extract forking method outside of VIP framework and update imports in fork tests
