+++ tests/hardhat/Fork/VRTVaultUpgrade.ts
@@ -4,7 +4,6 @@ import { expect } from "chai";
 import { BigNumber } from "ethers";
 import { ethers } from "hardhat";
 
-import { forking } from "../../../script/hardhat/fork/vip-framework";
 import {
   IAccessControlManagerV5__factory,
   VRT,
@@ -13,6 +12,7 @@ import {
   VRTVault__factory,
 } from "../../../typechain";
 import { IAccessControlManager } from "../../../typechain/contracts/Governance";
+import { forking } from "./utils";
 
 const FORK_MAINNET = process.env.FORK_MAINNET === "true";
 const bigNumber18 = BigNumber.from("1000000000000000000"); // 1e18
