+++ tests/hardhat/Fork/utils.ts
@@ -1,0 +1,24 @@
+import { network } from "hardhat";
+
+export const forking = (blockNumber: number, fn: () => void) => {
+  describe(`At block #${blockNumber}`, () => {
+    before(async () => {
+      await setForkBlock(blockNumber);
+    });
+    fn();
+  });
+};
+
+export async function setForkBlock(blockNumber: number) {
+  await network.provider.request({
+    method: "hardhat_reset",
+    params: [
+      {
+        forking: {
+          jsonRpcUrl: process.env.BSC_ARCHIVE_NODE,
+          blockNumber: blockNumber,
+        },
+      },
+    ],
+  });
+}
